confirm delete book ok

+++ app/src/app/components/book-item/book-details/book-details.component.ts
@@ -1,9 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
+import { MatDialog, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Book } from 'src/app/models/book';
 import { Chapter } from 'src/app/models/chapter';
 import { BookService } from 'src/app/services/book.service';
-import { ChapterService } from 'src/app/services/chapter.service';
+import { ConfirmComponent } from 'src/app/_helpers/confirm/confirm.component';
 
 @Component({
   selector: 'app-book-details',
@@ -11,39 +12,47 @@ import { ChapterService } from 'src/app/services/chapter.service';
   styleUrls: ['./book-details.component.scss']
 })
 export class BookDetailsComponent implements OnInit {
-
-  id:number;
-  book?:Book;
-  chapters:Chapter[];
-  
+  id: number;
+  book?: Book;
+  chapters: Chapter[];
+  @Inject(MAT_DIALOG_DATA) public data: any;
 
   constructor(private route: ActivatedRoute,
-    private bookService: BookService) { }
+    private bookService: BookService,
+    private matDialog: MatDialog) { }
 
   ngOnInit(): void {
     this.id = this.route.snapshot.params.id;
     try {
       this.bookService.getBookById(this.id).subscribe(
-        (dataBook:Book) => {
+        (dataBook: Book) => {
           this.book = dataBook;
           this.chapters = dataBook.chapter;
-          console.log(dataBook);
-          console.log(this.chapters);
-          
         }
       )
     } catch {
       console.log("__Error handled gracefully.")
     }
-    
+
   }
 
-  deleteBook(): void{
-    this.bookService.deleteBookById(this.id).subscribe(
-      (resp:any) => {
-        console.log('livre supprimé');
-      }
-    )
+  deleteBook(): void {
+    try {
+      const dialogRef = this.matDialog.open(ConfirmComponent, {
+        width: '350px',
+        data: "Voulez-vous supprimer ce livre ?"
+      });
+      dialogRef.afterClosed().subscribe(result => {
+        if (result) {
+          this.bookService.deleteBookById(this.id).subscribe(
+            (resp: any) => {
+              console.log('livre supprimé');
+            })
+        }
+      })
+    } catch {
+      console.log("__Error handled gracefully.")
+    }
   }
 
 }
